// backend/src/types/controlIva.ts
export interface SaleIva {
  cod_emp: number;
  llave: string;
  fecha: string;
  id_tipo_documento: string;
  documento: string;
  cod_cliente: string;
  gravadas_locales: number;
  gravadas_exportacion: number;
  ventas_exentas: number;
  ventas_no_sujetas: number;
  cuentas_a_terceros: number;
  rebajas_y_devoluciones: number;
  iva_retenido: number;
  iva_percibido: number;
  debito_fiscal: number;
  debito_fiscal_a_terceros: number;
  corr_maquina_registradora?: string;
  serie?: string;
  formulario_unico?: string;
  id_sucursal?: string;
  anulada?: number;
  es_rebajas_fac?: number;
  num_control?: string;
  nom_cliente?: string;
  registro_cliente?: string;
  nit_cliente?: string;
  nom_tipo_documento?: string;
}

export interface PaginatedResult<T> {
  data: T[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}